+++ src/containerapp-preview/HISTORY.rst
@@ -5,7 +5,7 @@ Release History
 upcoming
 ++++++
 * containerapp-preview requires the version of containerapp extension >= 0.3.36
-
+* `az containerapp create`: --yaml support property additionalPortMappings for api-version 2023-05-02-preview
 
 1.0.0b1
 ++++++